requirements.txt
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.4
pandas==2.2.3
pillow==11.0.0
pytz==2025.2
pywin32==304
sympy==1.13.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tzdata==2025.2
fastapi==0.115.12
uvicorn==0.34.0